requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
scikit-learn>=1.3.0
scipy>=1.10.0
Pillow>=9.5.0
opencv-python>=4.8.0
umap-learn>=0.5.3
jupyter>=1.0.0
ipython>=8.12.0
datasets>=2.14.0
tqdm>=4.65.0
diffusers==0.21.4
